refactor: move coroutines to app package
