+++ app/ws.py
@@ -1,0 +1,6 @@
+from domain.services.ws_client import WsClient
+
+
+async def ws_stream(ws: WsClient) -> None:
+    """Consume websocket stream indefinitely."""
+    await ws.stream()
